Add generic `From()` tests using base test class
Added a lot of tests to cover the `From()` method. Part of task #66.
